# examples/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/user/pageList": {
            "post": {
                "description": "获取用户分页列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "language（可选）",
                        "name": "language",
                        "in": "header"
                    },
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UserPageListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.UserPageListResp"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.UserPageListReq": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "v1.UserPageListResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}